requirements.txt
Package              Version
-------------------- ---------
asgiref              3.4.1
bleach               4.1.0
certifi              2021.10.8
Django               4.0
django-ckeditor      6.2.0
django-cors-headers  3.10.1
django-crispy-forms  1.13.0
django-debug-toolbar 3.2.4
django-filter        21.1
django-js-asset      1.2.2
django-summernote    0.8.20.0
django-taggit        2.0.0
django-tinymce4-lite 1.8.0
djangorestframework  3.13.1
jsmin                3.0.0
olefile              0.46
packaging            21.3
Pillow               8.4.0
pip                  21.2.4
psycopg2             2.9.2
psycopg2-binary      2.9.2
pyparsing            3.0.6
pytz                 2021.3
setuptools           58.0.4
six                  1.16.0
sqlparse             0.4.2
tzdata               2021.5
webencodings         0.5.1
wheel                0.37.0
wincertstore         0.2
